lib: fix sanitizer warnings for default_int_hash
int_hash.c:508:14: runtime error: left shift of 2572764 by 10 places cannot be represented in type 'int'
int_hash.c:510:14: runtime error: left shift of 2053447083 by 7 places cannot be represented in type 'int'
int_hash.c:510:6: runtime error: signed integer overflow: 848221568 + 2053447083 cannot be represented in type 'int'
int_hash.c:508:6: runtime error: signed integer overflow: 2118871040 + 64983770 cannot be represented in type 'int'
int_hash.c:506:14: runtime error: left shift of 158136035 by 4 places cannot be represented in type 'int'
int_hash.c:506:6: runtime error: signed integer overflow: 2090453200 + 130653325 cannot be represented in type 'int'
int_hash.c:504:14: runtime error: left shift of negative value -2147483647

+++ LIBRARY/common/int_hash.c
@@ -13,7 +13,7 @@
 
 static size_t g_num_of_collision;
 
-static size_t default_int_hash_function(int);
+static unsigned int default_int_hash_function(unsigned int);
 
 
 /*
@@ -498,8 +498,7 @@ int int_hash_iter_remove(INT_HASH_ITER *piter)
 /*
  *	derived from hashpjw, Dragon Book P436
  */
-
-static size_t default_int_hash_function(int key)
+static unsigned int default_int_hash_function(unsigned int key)
 {
 	key += (key << 12);
 	key ^= (key >> 22);
